Add check to see if target reached

+++ src/app/snake.rs
@@ -4,13 +4,13 @@ use crate::app::board::Direction;
 
 #[derive(Debug)]
 pub struct Snake {
-    points: Vec<(f64, f64)>
+    points: Vec<(i32, i32)>
 }
 
 impl Default for Snake {
     fn default() -> Self {
         Self {
-            points: vec![(2.0, 10.0)]
+            points: vec![(2, 10)]
         }
     }
 }
@@ -18,7 +18,7 @@ impl Default for Snake {
 impl Shape for Snake {
     fn draw(&self, painter: &mut Painter) {
         for (x, y) in self.points.iter() {
-            if let Some((x, y)) = painter.get_point(*x, *y) {
+            if let Some((x, y)) = painter.get_point(*x as f64, *y as f64) {
                 painter.paint(x, y, Color::Green);
             }
         }
@@ -26,17 +26,17 @@ impl Shape for Snake {
 }
 
 impl Snake {
-    pub fn head(&self) -> Option<&(f64, f64)> {
+    pub fn head(&self) -> Option<&(i32, i32)> {
         self.points.get(0)
     }
     pub fn move_snake(&mut self, direction: Direction) {
         let popped = self.points.pop();
         if let Some(popped) = popped {
             let delta = match direction {
-                Direction::Up => (0.0, 1.0),
-                Direction::Down => (0.0, -1.0),
-                Direction::Left => (-1.0, 0.0),
-                Direction::Right => (1.0, 0.0),
+                Direction::Up => (0, 1),
+                Direction::Down => (0, -1),
+                Direction::Left => (-1, 0),
+                Direction::Right => (1, 0),
             };
             self.points.push((popped.0 + delta.0, popped.1 + delta.1));
         }
